+++ source/CodeImatic.ObjectPascal.pas
@@ -5,11 +5,13 @@ interface
 uses
   System.SysUtils,  System.Classes, dwsCompiler,  NovusFileUtils,  dwsXPlatform,
   dwsComp, dwsExprs, dwsSymbols, dwsUtils, CodeImatic.Output, dwsUnitSymbols, System.IOUtils,
-  NovusObject, dwsDebugger, dwsDebugFunctions, dwsInfo,
+  NovusObject, dwsDebugger, dwsDebugFunctions, dwsInfo, dwsClasses, dwsByteBufferFunctions,
   dwsRTTIConnector, dwsRTTIFunctions, dwsScriptSource ,  CodeImatic.RuntimeBase,
-  dwsStringFunctions, dwsFunctions, CodeImatic.ObjectPascal.UnitBase,
+  dwsStringFunctions, dwsFunctions, CodeImatic.ObjectPascal.UnitBase, dwsVariantFunctions,
   dwsResultFunctions, dwsMathFunctions, dwsMathComplexFunctions, dwsFileFunctions,
-  CodeImatic.ObjectPascal.RTL.SystemExt;
+  CodeImatic.ObjectPascal.RTL.SystemExt, dwsTimeFunctions, dwsMath3DFunctions,
+  dwsGlobalVarsFunctions, dwsDataContext, CodeImatic.ObjectPascal.Symbol,
+  CodeImatic.ObjectPascal.Symbol.Byte;
 
 type
   tcimObjectPascal = class(TNovusobject)
@@ -24,6 +26,9 @@ type
     fsSearchPath: string;
     fDebugger: TdwsDebugger;
     fUnitList: tcimCustomUnitList;
+
+    procedure RegisterByteType(systemTable: TSystemSymbolTable);
+
 
     procedure DoDebugEval(exec: TdwsExecution; expr: TExprBase);
     procedure DoDebugSuspended(sender : TObject);
@@ -41,8 +46,9 @@ type
     function LoadStringFromFile(const FileName: string): string;
 
     procedure AddUnit(aUnit: tcimObjectPascalUnitBase);
+    procedure AddSymbol(aFuncSymbol: TFuncSymbol);
 
-    function Compile(aScript: String; aWorkingdirectory, aSearchPath: string; aCompileOnly: Boolean;aDebugger: boolean): boolean;
+    function Compile(aScript: String; aWorkingdirectory, aSearchPath: string; aCompileOnly: Boolean;aDebugger: boolean; aVerbose: boolean = false): boolean;
 
     property oCompiler: TDelphiWebScript
       read fCompiler;
@@ -96,28 +102,50 @@ begin
   fUnitList.AddUnit(aUnit);
 end;
 
-procedure tcimObjectPascal.AddUnits;
-Var
-  FCustomUnit: tdwsUnit;
-  FCustomFunction: TdwsFunction;
-  FCustomParameter: TdwsParameter;
+procedure tcimObjectPascal.AddSymbol(aFuncSymbol: TFuncSymbol);
 begin
-  AddUnit(tcimObjectPascalRTLSystem.Create(fCompiler, fRuntime));
+  (fCompiler.Config.SystemSymbols as TSymbolTable).AddSymbol(aFuncSymbol)
+end;
+
+procedure tcimObjectPascal.AddUnits;
+var
+  ByteSymbol: TBaseIntegerSymbol;
+begin
+  AddUnit(tcimObjectPascalRTLSystem.Create(fCompiler, fRuntime, nil));
+
+
+  TcimObjectPascalSymbolByte.RegisterSymbol(fCompiler.Config.SystemSymbols.SymbolTable);
+
+  //RegisterByteType(fCompiler.Config.SystemSymbols.SymbolTable);
+
+
+
 
 end;
 
 
-function tcimObjectPascal.Compile(aScript: String; aWorkingdirectory, aSearchPath: string; aCompileOnly: Boolean;aDebugger: boolean): boolean;
+function tcimObjectPascal.Compile(aScript: String; aWorkingdirectory,
+     aSearchPath: string;
+     aCompileOnly: Boolean;
+     aDebugger: boolean;
+     aVerbose: boolean = false): boolean;
 begin
   Result := false;
-  fsSearchPath := aSearchPath;
+
   fsWorkingdirectory := aWorkingdirectory;
+
+  fsSearchPath := aSearchPath;
 
   fCompiler.Config.OnNeedUnitEx := DoNeedUnitEx;
   fCompiler.Config.OnIncludeEx := DoIncludeEx;
 
   fCompiler.Config.ScriptPaths.Add(aWorkingdirectory);
   fCompiler.Config.ScriptPaths.Add(aSearchPath);
+
+  If aCompileOnly then
+     if (aVerbose) then fOutput.oLog.AddLogInformation('Compiling only ...')
+  else
+     if (aVerbose) then fOutput.oLog.AddLogInformation('Compiling ...');
 
   fProgram := fCompiler.Compile(aScript);
 
@@ -136,7 +164,11 @@ begin
     end;
 
   If Not aDebugger then
-    fExecute := fProgram.Execute
+   begin
+     if (aVerbose) then fOutput.oLog.AddLogInformation('Executing ...');
+
+     fExecute := fProgram.Execute
+   end
   else
   begin
     fDebugger.Breakpoints.Add(11, FProgram.SourceList[0].SourceFile.Name);
@@ -251,6 +283,25 @@ begin
    FDebugger.Resume;
   *)
 end;
+
+
+procedure tcimObjectPascal.RegisterByteType(systemTable: TSystemSymbolTable);
+var
+  byteTypeSym: TTypeSymbol;
+begin
+  // Check if Byte type already exists
+  var byteSymbol := systemTable.FindTypeLocal('Byte');
+  if byteSymbol = nil then
+  begin
+    // Create a new alias type to Integer
+   byteTypeSym := tcimObjectPascalSymbol.Create('Byte', systemTable.TypInteger);
+
+
+    // Optionally, you can later enhance this to enforce 0-255 range with custom casting
+    systemTable.AddSymbol(byteTypeSym);
+  end;
+end;
+